requirements.txt
aioredis==2.0.1
async-timeout==4.0.2
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
cleo==1.0.0a4
click==8.0.3
crashtest==0.3.1
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
Faker==12.1.0
Flask==2.0.2
hiredis==2.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pony==0.7.16
pptree==3.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pylev==1.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-ulid==1.0.3
redis==4.1.2
redis-om==0.0.17
requests==2.27.1
six==1.16.0
toml==0.10.2
types-redis==4.1.11
types-six==1.16.10
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
